Add files via upload
Today, with the development of technology, information processes are easily managed by computers. Large institutions, businesses, etc. A lot of data needs to be kept. Systems for keeping and managing data are frequently used today. Health care, which is one of the biggest needs of a country, is one of the areas where the most data should be kept and taken. One of the places where this system is used is hospitals. Hospital automation system is a system that should be used by all hospitals. Autonomous systems are used to register patients and doctors, and to ensure that patients benefit from health services properly. In this project, it is aimed to make the necessary program for the management of a hospital.



To use icons, change icon path from jpanel, button etc.